+++ tests/review.rs
@@ -115,6 +115,21 @@ fn test_simple() -> Result<(), std::io::Error> {
     Ok(())
 }
 
+#[test]
+fn test_generator_order() -> Result<(), std::io::Error> {
+    let dir = "test-resources/generator-order";
+    common::remove_sink(dir)?;
+    common::cmd(400)
+        .current_dir(dir)
+        .args(&["review", "generator-order"])
+        .assert()
+        .success()
+        .stdout("")
+        .stderr("");
+    common::check_sink(dir)?;
+    Ok(())
+}
+
 #[cfg(unix)]
 #[test]
 fn test_generator_url() -> Result<(), std::io::Error> {